fix(typescript/algokit_transact): use node & esm bundles for main & module
Initially the main and module of the package.json was set to wasm2js
under the assumption that only outdated software that didn't support
wasm would be using those fields over exports. This was an incorrect
assumption. This commit updates main & module to use the proper WASM
bundles but the wasm2js bundle is still defined under exports for
react-native