feat: indexer api client and minor jinja generator fixes (#228)
* feat: initial indexer client

* chore: client name generation fixes

* chore: regen indexer

* chore: fix msgpack detection in algod

* fix: handles top-level array schemas as type aliases

Updates the OAS parser to represent top-level array schemas as type aliases (e.g., Vec) instead of structs.

This simplifies the generated Rust code and avoids creating unnecessary empty structs for array types. Also it renames and refactors the 'StateDelta' and 'TealKeyValueStore' to be type aliases rather than structs.

* feat: Adds Indexer client to algokit-utils

Adds the Indexer client to algokit-utils, enabling interaction with the Algorand Indexer service.

This includes:
- Integration of the `indexer_client` crate.
- Updates to `AlgorandClient` to support Indexer configuration.
- Addition of methods in `ClientManager` to manage Indexer client instances.
- Adds pre-commit hooks for formatting, linting, and testing the API OAS generator code.
- Adds integration tests for the Indexer client.

* chore: wait for indexer helpers

* fix: add tweak to fix manual typos in indexer oas upstream descriptions; refine ci

Introduces a composite action to set up Rust, Bun, and Python dependencies.

Implements CI workflows for linting, testing, generating, and verifying the API client code. This includes output stability checks and compilation tests.

Adds a scheduled workflow to convert OpenAPI specifications and identify any necessary updates.

Fixes typos in the OpenAPI specification and generated code.

* refactor: adds indexer helpers for testing (aligning with ts utils)

This change introduces a set of helper functions to facilitate testing involving the Algorand Indexer.

It provides functionality to wait for the indexer to catch up with transactions or applications, with configurable retry logic and error handling.
This ensures tests can reliably verify the state of the indexer after submitting transactions.

This commit also removes the previous implementation of indexer waiting logic, and replaces it with the new helpers.

* chore: merge conflicts

* fix: adds support for string enums in OAS parser

Adds functionality to parse string enums from the OAS specification
and generate corresponding Rust enums. This includes adding a new
`enum_values` field to the `Schema` dataclass and a macro for
generating Rust enums from string values.

The changes also refactor the `OASParser` class to improve readability
and maintainability by extracting several methods for handling
vendor extensions, array schemas, and properties. This makes the code
easier to understand and modify in the future.

* chore: address pr comments